+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## 4.1.0 - 2022-0x-xx
 ### Added
 - Dynamic text variables for filter #186 => [wiki](https://github.com/Rello/analytics/wiki/Filter,-chart-options-&-drilldown#filter) for details
-- JSON data source: POST body; reading arrays #233 => [wiki](https://github.com/Rello/analytics/wiki/Datasource:-JSON) for details
 - Zoom in charts #246
+- Analysis functions (beta): Aggregate / disaggregate values #258
+- JSON data source: POST body; reading arrays #233 => [wiki](https://github.com/Rello/analytics/wiki/Datasource:-JSON) for details
 
 ### Changed
 - Cancel report load when another report is selected #249